Show line priority in line details page title
